backend/cv-processing-service/requirements.txt
ultralytics==8.0.186
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2 --index-url https://download.pytorch.org/whl/cpu
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6